bard_requirements.txt
# Automatically generated using PDM. Do not edit manually. 
requests
colorama
langdetect
deep_translator
google-cloud-translate
browser_cookie3